+++ src/translator/beam_search.h
@@ -20,7 +20,7 @@ private:
   Ptr<const Vocab> trgVocab_;
   Ptr<TensorAllocator> allocator_;
 
-  const float INVALID_PATH_SCORE = std::numeric_limits<float>::lowest(); // @TODO: observe this closely
+  const float INVALID_PATH_SCORE = -std::numeric_limits<float>::infinity(); // @TODO: observe this closely
   const bool PURGE_BATCH = true; // @TODO: diagnostic, to-be-removed once confirmed there are no issues.
 
 public: